revert(structure): move debugger out of annotations
